SI-2704 Handle aggregate signal query aliases (#34)
This allows one to, e.g., query for both the MIN and MAX odometer for each
day in an aggregate query. Before, only one of the aggregations would be computed
and it would be returned for each alias.

Instead of all aggregations grouping by timestamp and name, we now group by
the triplet timestamp, name, agg.